refactor: 캐러셀 컴포넌트와 그룹 버튼 컴포넌트 합침
