+++ src/pages/main/components/GroupCarousel.tsx
@@ -1,14 +1,15 @@
 import styled from '@emotion/styled';
+import { useNavigate } from 'react-router-dom';
 import Slider from 'react-slick';
 
 import '../styles/slick-theme.css';
 import '../styles/slick.css';
 
 import GroupContent from './GroupContent';
-import GroupNameButton from './GroupNameButton';
 
 import { groupPropTypes } from '../types/groupContent';
 
+import { MainIcnArrowPurple as MainIcnArrowPurpleIcon } from '../../../assets/svgs';
 import Spacing from '../../../components/commons/Spacing';
 
 export interface carouselItemPropTypes {
@@ -27,13 +28,21 @@ const GroupCarousel = ({ data }: carouselItemPropTypes) => {
     slidesToScroll: 1,
   };
 
+  const navigate = useNavigate();
+  const handleButtonOnClick = (groupId: string) => {
+    navigate(`/group/${groupId}`);
+  };
+
   return (
     <>
       {data?.map((moim) => (
         <CarouselWrapper key={moim.moimId}>
           <Spacing marginBottom="3.6" />
           <CarouselWithButtonLayout key={moim.moimId}>
-            <GroupNameButton groupName={moim.moimName} groupId={moim.moimId} />
+            <GroupButtonContainer onClick={() => handleButtonOnClick(moim.moimId)}>
+              {moim.moimName}
+              <MainIcnArrowPurpleIcon />
+            </GroupButtonContainer>
             <Spacing marginBottom="1.6" />
             <CarouselContainer>
               <CarouselBox {...settings} className="main">
@@ -68,6 +77,25 @@ const CarouselWrapper = styled.div`
 
 const CarouselWithButtonLayout = styled.div`
   margin-bottom: 3.2rem;
+`;
+
+const GroupButtonContainer = styled.button`
+  display: flex;
+  gap: 0.8rem;
+  align-items: center;
+  justify-content: flex-start;
+  padding: 0.6rem 1rem;
+
+  color: ${({ theme }) => theme.colors.mainViolet};
+
+  border: 1px solid ${({ theme }) => theme.colors.mainViolet};
+  border-radius: 0.8rem;
+
+  ${({ theme }) => theme.fonts.subtitle2};
+
+  &:hover {
+    background-color: ${({ theme }) => theme.colors.mileViolet};
+  }
 `;
 
 const CarouselContainer = styled.div`
